fix tasks

+++ task2.py
@@ -5,7 +5,7 @@ import numpy as np
 """
 
 arr = np.array([0, 1, 0, 2, 10, 0, 3]) # max -> 3
-nums = np.array([], dtype='int64')
+nums = np.array([], dtype='int32')
 for i in np.arange(1, np.size(arr)):
     if arr[i - 1] == 0:
         nums = np.append(nums, arr[i])
